requirements.txt
gTTS==1.1.8
playsound==1.2.1
SpeechRecognition==3.6.0
Pillow==4.0.0
cv2==1.0
numpy==1.12.0
progressbar2==3.12.0
